Fix mask not match object
