+++ Assets/AXIP/AILIA-MODELS/ObjectDetection/Yolov11Seg.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a137961e669fd44a5ab02ba97626da27
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
